+++ apps/web/registry/helpers/TextStyles.module.scss
@@ -1,0 +1,175 @@
+/* amino-ui-helper-file */
+
+.base {
+  display: flex;
+  align-items: center;
+  height: auto;
+  transition: var(--colorTransition);
+  color: inherit;
+
+  &--unselectable { user-select: none; }
+}
+
+.d1 {
+  font-family: var(--fontface_primary);
+  font-size: 42.5rem;
+}
+
+.d2 {
+  font-family: var(--fontface_primary);
+  font-size: 32rem;
+}
+
+.d3 {
+  font-family: var(--fontface_primary);
+  font-size: 24rem;
+}
+
+.d4 {
+  font-family: var(--fontface_primary);
+  font-size: 18rem;
+}
+
+.d5 {
+  font-family: var(--fontface_primary);
+  font-size: 13.4rem;
+}
+
+.d6 {
+  font-family: var(--fontface_primary);
+  font-size: 10rem;
+}
+
+.h1 {
+  font-family: var(--fontface_primary);
+  font-size: 3.8rem;
+}
+
+.h1-sm {
+  font-family: var(--fontface_primary);
+  font-size: 2.4rem;
+}
+
+.h2 {
+  font-family: var(--fontface_primary);
+  font-size: 2.4rem;
+}
+
+.h2-sm {
+  font-family: var(--fontface_primary);
+  font-size: 2.0rem;
+}
+
+.h3 {
+  font-family: var(--fontface_primary);
+  font-size: 2.0rem;
+}
+
+.h3-sm {
+  font-family: var(--fontface_primary);
+  font-size: 1.8rem;
+}
+
+.h4 {
+  font-family: var(--fontface_primary);
+  font-size: 1.8rem;
+}
+
+.h4-sm {
+  font-family: var(--fontface_primary);
+  font-size: 1.6rem;
+}
+
+.h5 {
+  font-family: var(--fontface_primary);
+  font-size: 1.6rem;
+}
+
+.h5-sm {
+  font-family: var(--fontface_primary);
+  font-size: 1.4rem;
+}
+
+.h6 {
+  font-family: var(--fontface_primary);
+  font-size: 1.4rem;
+}
+
+.h6-sm {
+  font-family: var(--fontface_primary);
+  font-size: 1.2rem;
+}
+
+.b1 {
+  font-family: var(--fontface_primary);
+  font-size: 7.6rem;
+}
+
+.b2 {
+  font-family: var(--fontface_primary);
+  font-size: 5.8rem;
+}
+
+.b3 {
+  font-family: var(--fontface_primary);
+  font-size: 4.2rem;
+}
+
+.b4 {
+  font-family: var(--fontface_primary);
+  font-size: 3.2rem;
+}
+
+.b5 {
+  font-family: var(--fontface_primary);
+  font-size: 2.4rem;
+}
+
+.b6 {
+  font-family: var(--fontface_primary);
+  font-size: 2.2rem;
+}
+
+.b7 {
+  font-family: var(--fontface_primary);
+  font-size: 2rem;
+}
+
+.b8 {
+  font-family: var(--fontface_primary);
+  font-size: 1.8rem;
+}
+
+.b9 {
+  font-family: var(--fontface_primary);
+  font-size: 1.6rem;
+}
+
+.b10 {
+  font-family: var(--fontface_primary);
+  font-size: 1.4rem;
+}
+
+.b11 {
+  font-family: var(--fontface_primary);
+  font-size: 1.2rem;
+}
+
+.b12 {
+  font-family: var(--fontface_primary);
+  font-size: 1.1rem;
+}
+
+.b13 {
+  font-family: var(--fontface_primary);
+  font-size: 1rem;
+}
+
+.fw-light { font-weight: light; }
+.fw-regular { font-weight: normal; }
+.fw-semibold { font-weight: semibold; }
+.fw-medium { font-weight: medium; }
+.fw-bold { font-weight: bold; }
+.fw-extrabold { font-weight: extrabold; }
+
+.fs-italic { font-style: italic; }
