+++ apps/web/registry/helpers/constants/geometry.ts
@@ -1,0 +1,10 @@
+// amino-ui-helper-file
+
+export const ORTHOGONAL = "orthogonal";
+export const ROUNDED = "rounded";
+export const ROUND = "round";
+
+export const AVAILABLE_CORNER_GEOMETRIES = [ORTHOGONAL, ROUNDED, ROUND];
+
+export const VERTICAL = "vertical";
+export const HORIZONTAL = "horizontal";
