+++ apps/web/registry/helpers/constants/theme.ts
@@ -1,0 +1,14 @@
+// amino-ui-helper-file
+
+export const THEME__PRIMARY = "primary";
+export const THEME__SECONDARY = "secondary";
+export const THEME__TERTIARY = "tertiary";
+export const THEME__QUATERNARY = "quaternary";
+export const THEME__QUINTENARY = "quintenary";
+
+export const AVAILABLE_THEMING_ORDERS = [
+  THEME__PRIMARY, THEME__SECONDARY, THEME__TERTIARY, THEME__QUATERNARY, THEME__QUINTENARY,
+];
+
+export const THEME__LIGHT = "light";
+export const THEME__DARK = "dark";
