+++ apps/web/registry/helpers/constants/ui.ts
@@ -1,0 +1,43 @@
+// amino-ui-helper-file
+
+export const VIEWBOX_CENTER_X = 50;
+export const VIEWBOX_CENTER_Y = 50;
+
+export const STROKE_LINECAP__BUTT = "butt";
+export const STROKE_LINECAP__ROUND = "round";
+export const STROKE_LINECAP__SQUARE = "square";
+export const AVAILABLE_STROKE_LINECAPS = [STROKE_LINECAP__BUTT, STROKE_LINECAP__ROUND, STROKE_LINECAP__SQUARE];
+
+export const POPOVER_PLACEMENT__BOTTOM = "bottom";
+export const POPOVER_PLACEMENT__BOTTOM_LEFT = "bottom left";
+export const POPOVER_PLACEMENT__BOTTOM_RIGHT = "bottom right";
+export const POPOVER_PLACEMENT__BOTTOM_START = "bottom start";
+export const POPOVER_PLACEMENT__BOTTOM_END = "bottom end";
+export const POPOVER_PLACEMENT__TOP = "top";
+export const POPOVER_PLACEMENT__TOP_LEFT = "top left";
+export const POPOVER_PLACEMENT__TOP_RIGHT = "top right";
+export const POPOVER_PLACEMENT__TOP_START = "top start";
+export const POPOVER_PLACEMENT__TOP_END = "top end";
+export const POPOVER_PLACEMENT__LEFT = "left";
+export const POPOVER_PLACEMENT__LEFT_TOP = "left top";
+export const POPOVER_PLACEMENT__LEFT_BOTTOM = "left bottom";
+export const POPOVER_PLACEMENT__START = "start";
+export const POPOVER_PLACEMENT__START_TOP = "start top";
+export const POPOVER_PLACEMENT__START_BOTTOM = "start bottom";
+export const POPOVER_PLACEMENT__RIGHT = "right";
+export const POPOVER_PLACEMENT__RIGHT_TOP = "right top";
+export const POPOVER_PLACEMENT__RIGHT_BOTTOM = "right bottom";
+export const POPOVER_PLACEMENT__END = "end";
+export const POPOVER_PLACEMENT__END_TOP = "end top";
+export const POPOVER_PLACEMENT__END_BOTTOM = "end bottom";
+
+export const AVAILABLE_POPOVER_PLACEMENT_POSITIONS = [
+  POPOVER_PLACEMENT__BOTTOM, POPOVER_PLACEMENT__BOTTOM_LEFT, POPOVER_PLACEMENT__BOTTOM_RIGHT,
+  POPOVER_PLACEMENT__BOTTOM_START, POPOVER_PLACEMENT__BOTTOM_END, POPOVER_PLACEMENT__TOP,
+  POPOVER_PLACEMENT__TOP_LEFT, POPOVER_PLACEMENT__TOP_RIGHT, POPOVER_PLACEMENT__TOP_START,
+  POPOVER_PLACEMENT__TOP_END, POPOVER_PLACEMENT__LEFT, POPOVER_PLACEMENT__LEFT_TOP,
+  POPOVER_PLACEMENT__LEFT_BOTTOM, POPOVER_PLACEMENT__START, POPOVER_PLACEMENT__START_TOP,
+  POPOVER_PLACEMENT__START_BOTTOM, POPOVER_PLACEMENT__RIGHT, POPOVER_PLACEMENT__RIGHT_TOP,
+  POPOVER_PLACEMENT__RIGHT_BOTTOM, POPOVER_PLACEMENT__END, POPOVER_PLACEMENT__END_TOP,
+  POPOVER_PLACEMENT__END_BOTTOM
+] as const;
