+++ apps/web/registry/helpers/types/accessibility.ts
@@ -1,0 +1,8 @@
+// amino-ui-helper-file
+
+export type TAriaLabelingProps = {
+  "aria-label"?: string,
+  "aria-labelledby"?: string,
+  "aria-describedby"?: string,
+  "aria-details"?: string
+}
