+++ apps/web/registry/helpers/types/data.ts
@@ -1,0 +1,3 @@
+// amino-ui-helper-file
+
+export type TGenericItem = { id: string; name: string; };
