+++ apps/web/registry/helpers/types/geometry.ts
@@ -1,0 +1,5 @@
+// amino-ui-helper-file
+
+import { AVAILABLE_CORNER_GEOMETRIES } from "@/src/constants/geometry";
+
+export type TAvailableCornerGeometries = typeof AVAILABLE_CORNER_GEOMETRIES[number];
