+++ apps/web/registry/helpers/types/theme.ts
@@ -1,0 +1,6 @@
+
+// amino-ui-helper-file
+
+import { AVAILABLE_THEMING_ORDERS } from "@src/constants/theme";
+
+export type TAvailableThemingOrders = typeof AVAILABLE_THEMING_ORDERS[number];
