+++ apps/web/registry/helpers/types/utils.ts
@@ -1,0 +1,7 @@
+// amino-ui-helper-file
+
+export type Enumerate<N extends number, Acc extends number[] = []> = Acc['length'] extends N
+  ? Acc[number]
+  : Enumerate<N, [...Acc, Acc['length']]>
+
+export type TIntRange<F extends number, T extends number> = Exclude<Enumerate<T>, Enumerate<F>>;
