+++ apps/web/registry/helpers/utils/data.ts
@@ -1,0 +1,4 @@
+// amino-ui-helper-file
+
+export const listBoxItemsListGen =
+  (items: string[]) => items.map((item: string, index) => ({ id: index.toString(), name: item }));
