+++ apps/web/registry/helpers/utils/ui.ts
@@ -1,8 +1,6 @@
-export const data = `export const listBoxItemsListGen =
-(items: string[]) => items.map((item: string, index) => ({ id: index.toString(), name: item }));
-`;
+// amino-ui-helper-file
 
-export const ui = `import { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 
 export const useIsDocumentHidden = () => {
   const [isDocumentHidden, setIsDocumentHidden] = useState(false);
@@ -15,4 +13,3 @@ export const useIsDocumentHidden = () => {
 
   return isDocumentHidden;
 };
-`;
